requirements.txt
scikit-learn==1.1.3
numpy==1.24.2
pandas==2.0.0
orange-canvas-core==0.1.30
orange-widget-base==4.20.0
Orange3==3.34.1